requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# GitHub Integration
PyGithub==2.1.1
python-dotenv==1.0.0

# Code Analysis
tree-sitter==0.20.4
tree-sitter-python==0.20.4
tree-sitter-javascript==0.20.1
semgrep==1.45.0
radon==6.0.1

# ML/AI Libraries
transformers==4.35.0
torch==2.1.0
sentence-transformers==2.2.2
scikit-learn==1.3.2

# Vector Database (for code similarity)
qdrant-client==1.7.0

# Graph Database (for dependencies)
neo4j==5.14.0

# Utilities
httpx==0.25.1
aiohttp==3.9.1
python-multipart==0.0.6
gitpython==3.1.40

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Caching
redis==5.0.1

# Code Quality
black==23.11.0
ruff==0.1.6
mypy==1.7.0



